Migrate old command implementation into new module structure
