+++ dsync-cli/src/command/file.rs
@@ -1,19 +1,87 @@
-pub(crate) async fn file_add() -> anyhow::Result<()> {
-    unimplemented!()
+use std::path::Path;
+
+use crate::command::utils;
+use anyhow::Context;
+use dsync_proto::cli::{AddFileRequest, ListLocalFilesRequest, client_api_client::ClientApiClient};
+
+use crate::command::model::LOOPBACK_ADDR_V4;
+
+use super::model::{GroupId, RemoteId};
+
+pub(crate) async fn file_add(file_path: impl AsRef<Path>) -> anyhow::Result<()> {
+    let file_path = file_path.as_ref();
+    let file_path_abs = match file_path.canonicalize() {
+        Ok(abs_path) => abs_path,
+        Err(err) => {
+            let message = format!(
+                "Failed to turn file_path: ${file_path:?} into absolute path with err: {err}"
+            );
+            log::error!("{message}");
+            return Err(anyhow::anyhow!(message));
+        }
+    };
+
+    let path_as_string = file_path_abs
+        .to_str()
+        .context("Looks like the specified path is not a valid unicode")?
+        .to_string();
+
+    let file_paths: Vec<String> = vec![path_as_string];
+    let request = tonic::Request::new(AddFileRequest {
+        file_path: file_paths,
+    });
+
+    let mut client = ClientApiClient::connect(LOOPBACK_ADDR_V4).await?;
+
+    log::info!("Sending request to server");
+    log::debug!("{request:?}");
+
+    let response = client.add_file(request).await?;
+
+    log::info!("Received response from server");
+    log::debug!("{response:?}");
+
+    anyhow::Ok(())
 }
 
 pub(crate) async fn file_remove() -> anyhow::Result<()> {
-    unimplemented!()
+    todo!()
 }
 
-pub(crate) async fn file_list() -> anyhow::Result<()> {
-    unimplemented!()
+pub(crate) async fn file_list(
+    remote_id: Option<RemoteId>,
+    group_id: Option<GroupId>,
+) -> anyhow::Result<()> {
+    if remote_id.is_some() {
+        todo!("Remote file list is not supported yet");
+    }
+
+    if group_id.is_some() {
+        todo!("Group file list is not supported yet");
+    }
+
+    let request = tonic::Request::new(ListLocalFilesRequest {});
+    let mut client = ClientApiClient::connect(LOOPBACK_ADDR_V4).await?;
+
+    log::info!("Sending request to server");
+    log::debug!("{request:?}");
+
+    let response = client.list_local_files(request).await?;
+
+    log::info!("Received response from server");
+    log::debug!("{response:?}");
+
+    let payload = response.into_inner();
+
+    utils::print_local_files_desc(&payload.file_descs);
+
+    anyhow::Ok(())
 }
 
 pub(crate) async fn file_sync() -> anyhow::Result<()> {
-    unimplemented!()
+    todo!()
 }
 
 pub(crate) async fn file_unsync() -> anyhow::Result<()> {
-    unimplemented!()
+    todo!()
 }
