+++ dsync-cli/src/command/group.rs
@@ -1,13 +1,13 @@
 use super::model::{GroupId, RemoteId};
 
 pub(crate) async fn group_create(_group_id: GroupId) -> anyhow::Result<()> {
-    unimplemented!()
+    todo!()
 }
 
 pub(crate) async fn group_delete(_group_id: GroupId) -> anyhow::Result<()> {
-    unimplemented!()
+    todo!()
 }
 
 pub(crate) async fn group_list(_remote_id: Option<RemoteId>) -> anyhow::Result<()> {
-    unimplemented!()
+    todo!()
 }
