+++ dsync-cli/src/command/model.rs
@@ -1,2 +1,4 @@
 pub(crate) type GroupId = String;
 pub(crate) type RemoteId = String;
+
+pub(super) const LOOPBACK_ADDR_V4: &str = "http://127.0.0.1:50051";
